+++ src/app/ContaPoupanca.java
@@ -6,7 +6,7 @@ public class ContaPoupanca extends ContaCorrente {
 	 */
 	private static final long serialVersionUID = 7870033380660416217L;
 
-	public ContaPoupanca(Cliente cliente, int senha) {
+	public ContaPoupanca(Cliente cliente, long senha) {
         super(cliente, senha);
     }
